refactor: Add id to TaskUiState.kt
