+++ app/src/main/java/com/london/tudee/presentation/model/TaskUiState.kt
@@ -1,9 +1,10 @@
-package com.london.tudee.presentation.components.task
+package com.london.tudee.presentation.model
 
 import androidx.annotation.DrawableRes
 import com.london.tudee.domain.entities.Priority
 
 data class TaskUiState(
+    val id: Int,
     val priority: Priority,
     @DrawableRes val iconResId: Int,
     val title: String,
